% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-cc-tools.R
\name{backConvertCrossNames}
\alias{backConvertCrossNames}
\title{backConvertCrossNames}
\usage{
backConvertCrossNames(crossnames, ...)
}
\arguments{
\item{...}{additional arguments}

\item{arg}{the argument, in the form of "CC001", "CC002", etc., or vector of these.}
}
\value{
Returns the number version of the CCXXX strain names in the vector.
}
\description{
This converts CC cross names.
}
\examples{
backConvertCrossNames(c("CC002", "CC001"))
}
