% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-cc-tools.R
\docType{package}
\name{cctools}
\alias{cctools}
\alias{cctools-package}
\title{cctools: A package for wrangling CC and CC-RIX related data.}
\description{
cctools: A package for wrangling CC and CC-RIX related data.
}
