% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-cc-tools.R
\name{colSwitch}
\alias{colSwitch}
\title{colSwitch}
\usage{
colSwitch(arg, ...)
}
\arguments{
\item{arg}{the argument, in the form of "AA", "BB", etc.}

\item{...}{additional arguments}
}
\value{
Returns standard CC founder colors in hex-code
}
\description{
Provide standard colors for CC founders.
}
\section{cc-tools functions}{

}

\examples{
colSwitch(c("A", "B", "C", "DD", "EE", "FF"))
}
