% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-cc-tools.R
\name{convertCrossNames}
\alias{convertCrossNames}
\title{convertCrossNames}
\usage{
convertCrossNames(crossnames, ...)
}
\arguments{
\item{...}{additional arguments}

\item{arg}{the argument, in the form of "8042", "8002", etc., or vector of these.}
}
\value{
Returns the CCXXX version of the numeric strain names in the vector.
}
\description{
This converts CC cross names.
}
\examples{
convertCrossNames(c("8042", "8002"))
}
