% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-cc-tools.R
\name{docBuildInstall}
\alias{docBuildInstall}
\alias{docBuildInstall}
\title{docBuildInstall}
\usage{
docBuildInstall(dir, ...)

docBuildInstall(dir, ...)
}
\arguments{
\item{dir}{the directory of the R package to be documented, built, installed}

\item{...}{additional arguments}

\item{dir}{the directory path}

\item{...}{additional arguments}
}
\value{
Returns string without dot/underscore.

Installs the R package.
}
\description{
This documents, builds, and installs an R package based on its directory.

Document, build, and install an R package from its decompressed directory.
}
\examples{
# Not Run
# Not Run
}
