% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-cc-tools.R
\name{lettersToNumbers}
\alias{lettersToNumbers}
\title{lettersToNumbers}
\usage{
lettersToNumbers(arg, ...)
}
\arguments{
\item{arg}{the argument, in the form of "A", "B", ... "H", or vector of these.}

\item{...}{additional arguments}
}
\value{
Returns standard CC founder numbers from single letters.
}
\description{
This converts strain name letters to numbers.
}
\examples{
lettersToNumbers(c("A", "B", "C", "D", "E", "F"))
}
