% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-cc-tools.R
\name{lettersToStrains}
\alias{lettersToStrains}
\title{lettersToStrains}
\usage{
lettersToStrains(arg, ...)
}
\arguments{
\item{arg}{the argument, in the form of "AJ", "B6", etc.}

\item{...}{additional arguments}
}
\value{
Returns standard CC short strain names from letter codes.
}
\description{
Convert CC letter names to short strain names.
}
\examples{
lettersToStrains(c("AJ", "B6"))
}
