% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-cc-tools.R
\name{parseDamSire}
\alias{parseDamSire}
\title{parseDamSire}
\usage{
parseDamSire(data.frame, cross.colname, ...)
}
\arguments{
\item{data.frame}{the data frame with dam x sire names}

\item{cross.colname}{the column name with dam x sire}

\item{...}{additional arguments}
}
\value{
Returns string without dot/underscore.
}
\description{
This separates damxsire into dam and sire.
}
\examples{
# Not Run
}
