% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-cc-tools.R
\docType{package}
\name{PLMcctools}
\alias{PLMcctools}
\alias{PLMcctools-package}
\title{PLMcctools: A package for wrangling CC and CC-RIX related data.}
\description{
PLMcctools: A package for wrangling CC and CC-RIX related data.
}
