% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-cc-tools.R
\name{prob.heatmap.sdp}
\alias{prob.heatmap.sdp}
\title{prob.heatmap.sdp}
\usage{
prob.heatmap.sdp(marker, p.value = NULL, genomecache, model = "additive",
  phenotype, phenotype.data, merge.by = "SUBJECT.NAME",
  founder.labels = NULL, founder.cex = 1, founder.line = 1,
  founder.col = rep("black", length(levels(as.factor(sdp)))),
  phenotype.lab.cex = 1, phenotype.num.cex = 1, phenotype.num.padj = NA,
  phenotype.line = NA, phenotype.num.line = NA, include.ramp = TRUE,
  ramp.label.cex = 0.7, ramp.label.line = 0.5, prob.axis.cex = 1,
  include.marker = TRUE, marker.line = 1,
  alternative.phenotype.label = NULL, alternative.marker.label = NULL,
  sdp = NULL)
}
\arguments{
\item{...}{additional arguments}
}
\value{
Returns rescaled palette, with same length of colors as original scale
}
\description{
This is an SDP-collapsed version of the miqtl::prob.heatmap function.
}
\examples{
# Not Run
}
