% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-cc-tools.R
\name{removeDots}
\alias{removeDots}
\title{removeDots}
\usage{
removeDots(x, ...)
}
\arguments{
\item{x}{the dot-containing string
or a vector of character strings}

\item{...}{additional arguments}
}
\value{
Returns string without dot/underscore.
}
\description{
This removes dots (and underscores) from a string, replacing with a space.
}
\examples{
removeDots("This_is_a.test.string.")
}
