% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-cc-tools.R
\name{rescale.palette}
\alias{rescale.palette}
\title{rescale.palette}
\usage{
rescale.palette(plot.range, data.range, color.len = 1000,
  start.pal = colorRampPalette(c("white", "black"))(n = color.len), ...)
}
\arguments{
\item{plot.range}{the range of values for plotting (i.e., range of strain means)}

\item{data.range}{the range of values for the color ramp scale (i.e., range of strain means across all plots)}

\item{start.pal}{the starting palette}

\item{...}{additional arguments}
}
\value{
Returns rescaled palette, with same length of colors as original scale
}
\description{
This rescales the palette based on where the plot range falls within the data range.
}
\examples{
# Not Run
}
