% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-cc-tools.R
\name{strainsShortToFull}
\alias{strainsShortToFull}
\title{strainsShortToFull}
\usage{
strainsShortToFull(arg, ...)
}
\arguments{
\item{arg}{the argument, in the form of "AJ", "B6", etc., or a vector of these.}

\item{...}{additional arguments}
}
\value{
Returns standard CC founder letters from strain names.
}
\description{
This converts short strainnames to full strain names.
}
\examples{
strains <- c("AJ", "B6", "129", "NOD", "NZO", "CAST", "PWK", "WSB")
strainsShortToFull(strains)
}
