% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-cc-tools.R
\name{strainsToLetters}
\alias{strainsToLetters}
\title{strainsToLetters}
\usage{
strainsToLetters(arg, ...)
}
\arguments{
\item{arg}{the argument, in the form of "AJ", "C57BL6J", etc., or a vector of these.}

\item{...}{additional arguments}
}
\value{
Returns standard CC founder letters from strain names.
}
\description{
This converts strain names to strain letters.
}
\examples{
strains <- c("A_J", "C57BL6J", "129S1_SvImJ", "NOD_ShiLtJ", "NZO_HlLtJ", "CAST_EiJ", "PWK_PhJ", "WSB_EiJ")
strainsToLetters(strains)
}
