% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm-cc-tools.R
\name{underNumAlpha}
\alias{underNumAlpha}
\title{underNumAlpha}
\usage{
underNumAlpha(ids, ...)
}
\arguments{
\item{...}{additional arguments}

\item{arg}{the argument which has numeric and character strings next to each ther
or a vector of character strings}
}
\value{
Returns elements with underscores between letters and numbers
}
\description{
This adds underscores between alphabet and numeric characters
}
\examples{
underNumAlpha(c("135RL","136R"))
}
